build for app gh 21/4/2022

+++ lib/models/address.dart
@@ -2,7 +2,7 @@ class Address {
   Address({this.heading, this.lat, this.lng, this.address});
 
   String? heading;
-  int? lat;
-  int? lng;
+  double? lat;
+  double? lng;
   String? address;
 }
